+++ backend/src/main/java/linklibrary/controller/PostController.java
@@ -1,9 +1,11 @@
 package linklibrary.controller;
 
-import io.swagger.annotations.ApiImplicitParam;
-import io.swagger.annotations.ApiImplicitParams;
 import io.swagger.annotations.ApiOperation;
-import linklibrary.dto.*;
+import linklibrary.dto.request.PostFormDto;
+import linklibrary.dto.response.CategoryDto;
+import linklibrary.dto.response.MainPageDto;
+import linklibrary.dto.response.PostDto;
+import linklibrary.dto.response.ResponseData;
 import linklibrary.security.auth.PrincipalDetails;
 import linklibrary.service.CategoryService;
 import linklibrary.service.PostService;
