--- backend/src/main/java/linklibrary/dto/CategoryDto.java
@@ -1,18 +1,0 @@
-package linklibrary.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import javax.validation.constraints.NotBlank;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class CategoryDto {
-
-    private Long categoryId;
-    @NotBlank(message = "카테고리명을 입력해주세요.")
-    private String name;
-
-}
