--- backend/src/main/java/linklibrary/dto/CategoryFormDto.java
@@ -1,13 +1,0 @@
-package linklibrary.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Getter @Setter
-@AllArgsConstructor
-@NoArgsConstructor
-public class CategoryFormDto {
-    private String name;
-}
