--- backend/src/main/java/linklibrary/dto/ResponseData.java
@@ -1,12 +1,0 @@
-package linklibrary.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-
-@Data
-@AllArgsConstructor
-public class ResponseData {
-    private String message;
-    private Object data;
-
-}
