--- backend/src/main/java/linklibrary/dto/UserDto.java
@@ -1,16 +1,0 @@
-package linklibrary.dto;
-
-import linklibrary.entity.Role;
-import lombok.*;
-
-@Getter
-@Setter
-@AllArgsConstructor
-@NoArgsConstructor
-@Builder
-public class UserDto {
-    private Long userId;
-    private String loginId;
-    private String password;
-    private Role role;
-}
