+++ backend/src/main/java/linklibrary/dto/request/CategoryFormDto.java
@@ -1,0 +1,19 @@
+package linklibrary.dto.request;
+
+import io.swagger.v3.oas.annotations.media.Schema;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter @Setter
+@AllArgsConstructor
+@NoArgsConstructor
+@Schema(title = "카테고리 작명", description = "카테고리 작명 하기")
+public class CategoryFormDto {
+    //@Schema 어노테이션은 Swagger 문서에서 해당 객체에 대한 정보를 제공
+    //    @ApiModelProperty(example = "abcde1") 이거 로그인폼DTO에 있는데 이거랑 같은 기능인듯 ?
+    //찾아봤더니 @Schema가 최신 swagger버전에 어울린다함.
+    @Schema(title = "제목", example = "영화")
+    private String name;
+}
