+++ backend/src/main/java/linklibrary/dto/request/PostFormDto.java
@@ -1,4 +1,4 @@
-package linklibrary.dto;
+package linklibrary.dto.request;
 
 import linklibrary.entity.Category;
 import linklibrary.entity.User;
@@ -18,7 +18,9 @@ import java.util.Date;
 public class PostFormDto { //포스트 생성 폼
 
     //https://sanghye.tistory.com/36 validation 관련 블로그
-
+//@Schema 어노테이션은 Swagger 문서에서 해당 객체에 대한 정보를 제공
+    //    @ApiModelProperty(example = "abcde1") 이거 로그인폼DTO에 있는데 이거랑 같은 기능인듯 ?
+    //찾아봤더니 @Schema가 최신 swagger버전에 어울린다함.
     private Long postId;
 
     @NotBlank(message = "제목을 입력해주세요")
