+++ backend/src/main/java/linklibrary/dto/request/ValidateIdForm.java
@@ -1,4 +1,4 @@
-package linklibrary.dto;
+package linklibrary.dto.request;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -11,6 +11,9 @@ import javax.validation.constraints.Pattern;
 @AllArgsConstructor
 @NoArgsConstructor //기본 생성자가 없어서 나는 오류였어요. 수정했습니다
 public class ValidateIdForm {
+    //@Schema 어노테이션은 Swagger 문서에서 해당 객체에 대한 정보를 제공
+    //    @ApiModelProperty(example = "abcde1") 이거 로그인폼DTO에 있는데 이거랑 같은 기능인듯 ?
+    //찾아봤더니 @Schema가 최신 swagger버전에 어울린다함.
     @NotBlank(message = "아이디는 필수 입력 값입니다.")
     @Pattern(regexp = "^[a-z0-9]{4,20}$", message = "아이디는 영어 소문자와 숫자만 사용하여 4~20자리여야 합니다.")
     private String loginId;
