+++ backend/src/main/java/linklibrary/dto/response/MainPageDto.java
@@ -1,0 +1,26 @@
+package linklibrary.dto.response;
+
+import io.swagger.v3.oas.annotations.media.Schema;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+import org.springframework.data.domain.Page;
+
+import java.util.List;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+@Builder
+@Schema(title = "메인 페이지 요청", description = "카테고리 리스트,게시글 리스트 및 전체 수,현재 카테고리 요청 ")
+public class MainPageDto {
+    //@Schema 어노테이션은 Swagger 문서에서 해당 객체에 대한 정보를 제공
+    //    @ApiModelProperty(example = "abcde1") 이거 로그인폼DTO에 있는데 이거랑 같은 기능인듯 ?
+    //찾아봤더니 @Schema가 최신 swagger버전에 어울린다함.
+
+    private List<CategoryDto> categoryDtoList;
+    private Page<PostDto1> postDtoList;
+    private Long total;
+    private String currentCategory;
+}
