+++ backend/src/main/java/linklibrary/dto/response/PostDto1.java
@@ -1,4 +1,4 @@
-package linklibrary.dto;
+package linklibrary.dto.response;
 
 import com.querydsl.core.annotations.QueryProjection;
 import lombok.AllArgsConstructor;
@@ -13,6 +13,9 @@ import java.time.LocalDateTime;
 @NoArgsConstructor
 @Builder  //조회용
 public class PostDto1 {
+    //@Schema 어노테이션은 Swagger 문서에서 해당 객체에 대한 정보를 제공
+    //    @ApiModelProperty(example = "abcde1") 이거 로그인폼DTO에 있는데 이거랑 같은 기능인듯 ?
+    //찾아봤더니 @Schema가 최신 swagger버전에 어울린다함.
     private Long postId;
     private String title;
     @Lob
