+++ backend/src/main/java/linklibrary/dto/response/ResponseData.java
@@ -1,0 +1,16 @@
+package linklibrary.dto.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+@Data
+@AllArgsConstructor
+public class ResponseData {
+    //@Schema 어노테이션은 Swagger 문서에서 해당 객체에 대한 정보를 제공
+    //    @ApiModelProperty(example = "abcde1") 이거 로그인폼DTO에 있는데 이거랑 같은 기능인듯 ?
+    //찾아봤더니 @Schema가 최신 swagger버전에 어울린다함.
+
+    private String message;
+    private Object data;
+
+}
