+++ backend/src/main/java/linklibrary/exception/ExceptionController.java
@@ -1,6 +1,6 @@
 package linklibrary.exception;
 
-import linklibrary.dto.ResponseData;
+import linklibrary.dto.response.ResponseData;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
