+++ backend/src/main/java/linklibrary/repository/PostRepositoryCustom.java
@@ -1,11 +1,8 @@
 package linklibrary.repository;
 
-import linklibrary.dto.PostDto;
-import linklibrary.dto.PostDto1;
+import linklibrary.dto.response.PostDto1;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
-
-import java.util.List;
 
 public interface PostRepositoryCustom {
     Page<PostDto1> findPostDtos(Long userId, String bookmark, String sort, String keyword, Long categoryId, Pageable pageable);
