+++ backend/src/main/java/linklibrary/repository/PostRepositoryImpl.java
@@ -3,9 +3,8 @@ package linklibrary.repository;
 import com.querydsl.core.types.OrderSpecifier;
 import com.querydsl.core.types.dsl.BooleanExpression;
 import com.querydsl.jpa.impl.JPAQueryFactory;
-import linklibrary.dto.PostDto1;
-import linklibrary.dto.QPostDto1;
-import linklibrary.entity.QCategory;
+import linklibrary.dto.response.PostDto1;
+import linklibrary.dto.response.QPostDto1;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.Pageable;
