+++ backend/src/test/java/linklibrary/service/PostServiceTest.java
@@ -1,6 +1,6 @@
 package linklibrary.service;
 
-import linklibrary.dto.PostFormDto;
+import linklibrary.dto.request.PostFormDto;
 import linklibrary.entity.Category;
 import linklibrary.entity.Post;
 import linklibrary.entity.Role;
@@ -8,9 +8,7 @@ import linklibrary.entity.User;
 import linklibrary.repository.PostRepository;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.transaction.annotation.Transactional;
 
 import javax.persistence.EntityManager;
